# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SIP Agent Configuration Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DashboardStatsExtensionsByProviderItem = {
  provider: string;
  count: number;
};

export interface DashboardStats {
  totalClients: number;
  totalExtensions: number;
  totalAgentConfigs: number;
  extensionsByProvider: DashboardStatsExtensionsByProviderItem[];
}

export interface Client {
  id: number;
  name: string;
  description?: string | null;
  sipDomain?: string | null;
  sipServer?: string | null;
  /** Yeastar PBX HTTP API base URL for outbound call triggering */
  yeastarApiUrl?: string | null;
  /** Yeastar API token */
  yeastarApiToken?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateClientInput {
  name: string;
  description?: string | null;
  sipDomain?: string | null;
  sipServer?: string | null;
  yeastarApiUrl?: string | null;
  yeastarApiToken?: string | null;
}

export type AiProvider = typeof AiProvider[keyof typeof AiProvider];


export const AiProvider = {
  openai: 'openai',
  elevenlabs: 'elevenlabs',
  gemini: 'gemini',
  deepgram: 'deepgram',
  cartesia: 'cartesia',
} as const;

export type AgentConfigMode = typeof AgentConfigMode[keyof typeof AgentConfigMode];


export const AgentConfigMode = {
  inbound: 'inbound',
  outbound: 'outbound',
  both: 'both',
} as const;

export interface AgentConfig {
  id: number;
  name: string;
  provider: AiProvider;
  apiKey: string;
  voiceId?: string | null;
  modelId?: string | null;
  systemPrompt?: string | null;
  greeting?: string | null;
  mode?: AgentConfigMode;
  language?: string | null;
  /** JSON string for provider-specific extra fields */
  extraConfig?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Extension {
  id: number;
  clientId?: number | null;
  agentConfigId?: number | null;
  extensionNumber: string;
  displayName?: string | null;
  sipUsername: string;
  sipAuthId: string;
  sipPassword: string;
  createdAt: string;
  updatedAt: string;
  client?: Client | null;
  agentConfig?: AgentConfig | null;
}

export interface CreateExtensionInput {
  clientId?: number | null;
  agentConfigId?: number | null;
  extensionNumber: string;
  displayName?: string | null;
  sipUsername: string;
  sipAuthId: string;
  sipPassword: string;
}

export interface CreateAgentConfigInput {
  name: string;
  provider: AiProvider;
  apiKey: string;
  voiceId?: string | null;
  modelId?: string | null;
  systemPrompt?: string | null;
  greeting?: string | null;
  mode?: AgentConfigMode;
  language?: string | null;
  extraConfig?: string | null;
}

export type AgentToolExecutionType = typeof AgentToolExecutionType[keyof typeof AgentToolExecutionType];


export const AgentToolExecutionType = {
  http_request: 'http_request',
  webhook: 'webhook',
  transfer_call: 'transfer_call',
  hang_up: 'hang_up',
  send_dtmf: 'send_dtmf',
  custom_js: 'custom_js',
} as const;

export interface AgentTool {
  id: number;
  agentConfigId: number;
  name: string;
  description: string;
  /** JSON Schema string describing the tool's parameters */
  parametersSchema?: string | null;
  executionType: AgentToolExecutionType;
  /** JSON config specific to the execution type */
  executionConfig?: string | null;
  /** Execution timeout in seconds */
  timeout: number;
  requireConfirmation: boolean;
  enabled: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAgentToolInput {
  agentConfigId: number;
  name: string;
  description: string;
  parametersSchema?: string | null;
  executionType: AgentToolExecutionType;
  executionConfig?: string | null;
  timeout?: number;
  requireConfirmation?: boolean;
  enabled?: boolean;
  sortOrder?: number;
}

export type OutboundCallStatus = typeof OutboundCallStatus[keyof typeof OutboundCallStatus];


export const OutboundCallStatus = {
  pending: 'pending',
  dialing: 'dialing',
  active: 'active',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface OutboundCall {
  id: number;
  extensionId?: number | null;
  phoneNumber: string;
  callerId?: string | null;
  /** JSON string of runtime variables */
  variables?: string | null;
  firstMessage?: string | null;
  systemPromptOverride?: string | null;
  /** JSON string of caller metadata */
  metadata?: string | null;
  webhookUrl?: string | null;
  status: OutboundCallStatus;
  callId?: string | null;
  /** JSON string of call outcome */
  result?: string | null;
  error?: string | null;
  createdAt: string;
  updatedAt: string;
}

/**
 * Runtime context variables injected into the AI conversation
 */
export type TriggerOutboundCallInputVariables = { [key: string]: unknown } | null;

/**
 * Arbitrary metadata stored with the call record
 */
export type TriggerOutboundCallInputMetadata = { [key: string]: unknown } | null;

export interface TriggerOutboundCallInput {
  /** Extension to use for the outbound call */
  extensionId: number;
  /** Target phone number to call */
  phoneNumber: string;
  /** Caller ID to present to the callee */
  callerId?: string | null;
  /** Runtime context variables injected into the AI conversation */
  variables?: TriggerOutboundCallInputVariables;
  /** Override the agent's first message for this call */
  firstMessage?: string | null;
  /** Override the agent's system prompt for this call */
  systemPromptOverride?: string | null;
  /** Arbitrary metadata stored with the call record */
  metadata?: TriggerOutboundCallInputMetadata;
  /** URL to POST call result to when the call ends */
  webhookUrl?: string | null;
}

export type OutboundContextVariables = { [key: string]: unknown } | null;

export interface OutboundContext {
  pending: boolean;
  firstMessage?: string | null;
  systemPromptOverride?: string | null;
  variables?: OutboundContextVariables;
  callId?: number | null;
}

export type ListExtensionsParams = {
clientId?: number;
};

export type ListAgentToolsParams = {
agentConfigId: number;
};

export type ListOutboundCallsParams = {
extensionId?: number;
limit?: number;
};