# lib/api-zod/src/generated/types/agentTool.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SIP Agent Configuration Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentToolExecutionType } from './agentToolExecutionType';

export interface AgentTool {
  id: number;
  agentConfigId: number;
  name: string;
  description: string;
  /** JSON Schema string describing the tool's parameters */
  parametersSchema?: string | null;
  executionType: AgentToolExecutionType;
  /** JSON config specific to the execution type */
  executionConfig?: string | null;
  /** Execution timeout in seconds */
  timeout: number;
  requireConfirmation: boolean;
  enabled: boolean;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}