# lib/api-zod/src/generated/types/agentToolExecutionType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SIP Agent Configuration Manager API
 * OpenAPI spec version: 0.1.0
 */

export type AgentToolExecutionType = typeof AgentToolExecutionType[keyof typeof AgentToolExecutionType];


export const AgentToolExecutionType = {
  http_request: 'http_request',
  webhook: 'webhook',
  transfer_call: 'transfer_call',
  hang_up: 'hang_up',
  send_dtmf: 'send_dtmf',
  custom_js: 'custom_js',
} as const;