# lib/api-zod/src/generated/types/outboundCall.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SIP Agent Configuration Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { OutboundCallStatus } from './outboundCallStatus';

export interface OutboundCall {
  id: number;
  extensionId?: number | null;
  phoneNumber: string;
  callerId?: string | null;
  /** JSON string of runtime variables */
  variables?: string | null;
  firstMessage?: string | null;
  systemPromptOverride?: string | null;
  /** JSON string of caller metadata */
  metadata?: string | null;
  webhookUrl?: string | null;
  status: OutboundCallStatus;
  callId?: string | null;
  /** JSON string of call outcome */
  result?: string | null;
  error?: string | null;
  createdAt: Date;
  updatedAt: Date;
}