# lib/api-zod/src/generated/types/outboundCallStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SIP Agent Configuration Manager API
 * OpenAPI spec version: 0.1.0
 */

export type OutboundCallStatus = typeof OutboundCallStatus[keyof typeof OutboundCallStatus];


export const OutboundCallStatus = {
  pending: 'pending',
  dialing: 'dialing',
  active: 'active',
  completed: 'completed',
  failed: 'failed',
} as const;