# lib/api-zod/src/generated/types/outboundContext.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SIP Agent Configuration Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { OutboundContextVariables } from './outboundContextVariables';

export interface OutboundContext {
  pending: boolean;
  firstMessage?: string | null;
  systemPromptOverride?: string | null;
  variables?: OutboundContextVariables;
  callId?: number | null;
}